+++ __tests__/execute-action.ts
@@ -31,12 +31,10 @@ test('When action returns function execute it', () => {
     expect(result instanceof Rx.Observable).toEqual(true)
 });
 
-test('When action returns empty array return event', () => {
-    expect(() => {
-        executeAction(event => {
-            return []
-        }, null, null)
-    }).toThrow();
+test('When action returns empty array do nothing', () => {
+    executeAction(event => {
+        return []
+    }, null, null)
 });
 
 test('When action returns array execute in order', () => {