+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Exposes `matches` function used for pattern matching logic in Eventline to enable ability to add own custom patterns.
 - Removes dependencies which were installed but were only needed for development of this package and not at runtime.
+- You can now return an empty array of actions to indicate you don't want anything to happen
 
 # v0.2.18